+++ src/main/java/com/wlgzs/huipeasant/entity/User.java
@@ -15,10 +15,10 @@ import java.io.Serializable;
     public class User implements Serializable {
         @Id
         @GeneratedValue(strategy = GenerationType.IDENTITY)
-        private Integer userId;                          //用户ID
+        private long userId;                          //用户ID
         private String  phoneNumber;                     //用户手机号
         private String password;                        //密码
         private String headPortrait;                    //头像地址
         private String  nickName;                       // 昵称
-        private Integer roleId;                         //角色id
+        private long roleId;                         //角色id
     }
